requirements.txt
streamlit>=1.28.0
langchain>=0.0.340
langchain-community>=0.0.20
gitpython>=3.1.40
chromadb>=0.4.18
sentence-transformers>=2.2.2
groq>=0.4.1
python-dotenv>=1.0.0
tiktoken>=0.5.2
plotly>=5.17.0
pandas>=2.1.4
huggingface-hub>=0.16.0
transformers>=4.30.0
torch>=1.13.0
numpy>=1.21.0